+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 3.4.8 (unreleased)
 ------------------
 
+- Add new geolocated faceted view
+  [boulch]
+
 - Add geolocation field to project content type
   [boulch]